fix: only show focus style if not in popover
